Move MainAsset classes to their own directory
Split Printer
