+++ src/Glpi/Inventory/Asset/Printer.php
@@ -9,7 +9,6 @@
  *
  * @copyright 2015-2024 Teclib' and contributors.
  * @copyright 2003-2014 by the INDEPNET Development Team.
- * @copyright 2010-2022 by the FusionInventory Development Team.
  * @licence   https://www.gnu.org/licenses/gpl-3.0.html
  *
  * ---------------------------------------------------------------------
@@ -36,75 +35,25 @@
 
 namespace Glpi\Inventory\Asset;
 
-use Blacklist;
-use CommonDBTM;
+use AutoUpdateSystem;
 use Glpi\Asset\Asset_PeripheralAsset;
 use Glpi\Inventory\Conf;
-use IPAddress;
-use Printer as GPrinter;
-use PrinterLog;
-use PrinterModel;
-use PrinterType;
 use RuleDictionnaryPrinterCollection;
+use Printer as GPrinter;
 use RuleImportAssetCollection;
 
-class Printer extends NetworkEquipment
+class Printer extends InventoryAsset
 {
-    private $counters;
-
-    public function __construct(CommonDBTM $item, $data)
-    {
-        $this->extra_data['pagecounters'] = null;
-        parent::__construct($item, $data);
-    }
-
-    protected function getModelsFieldName(): string
-    {
-        return PrinterModel::getForeignKeyField();
-    }
-
-    protected function getTypesFieldName(): string
-    {
-        return PrinterType::getForeignKeyField();
-    }
-
     public function prepare(): array
     {
-        if ($this->item->getType() != GPrinter::getType() && $this->conf->import_printer == 0) {
-            return [];
-        }
-
-        parent::prepare();
-
-        if (!property_exists($this->raw_data->content ?? new \stdClass(), 'network_device')) {
-            $autoupdatesystems_id = $this->data[0]->autoupdatesystems_id;
-            $this->data = [];
-            foreach ($this->raw_data as $val) {
-                $val->autoupdatesystems_id = $autoupdatesystems_id;
-                $val->last_inventory_update = $_SESSION['glpi_currenttime'];
-                $this->data[] = $val;
-            }
-        }
-
         $rulecollection = new RuleDictionnaryPrinterCollection();
 
-        $mapping_pcounter = [
-            'total'        => 'total_pages',
-            'black'        => 'bw_pages',
-            'color'        => 'color_pages',
-            'duplex'       => 'rv_pages', //keep first, rectoverso is the standard and should be used if present
-            'rectoverso'   => 'rv_pages',
-            'scanned'      => 'scanned',
-            'printtotal'   => 'prints',
-            'printblack'   => 'bw_prints',
-            'printcolor'   => 'color_prints',
-            'copytotal'    => 'copies',
-            'copyblack'    => 'bw_copies',
-            'copycolor'    => 'color_copies',
-            'faxtotal'     => 'faxed',
-        ];
-
         foreach ($this->data as $k => &$val) {
+            //set update system
+            $val->autoupdatesystems_id = AutoUpdateSystem::NATIVE_INVENTORY;
+            $val->last_inventory_update = $_SESSION["glpi_currenttime"];
+            $val->is_deleted = 0;
+
             if (property_exists($val, 'port') && strstr($val->port, "USB")) {
                 $val->have_usb = 1;
             } else {
@@ -112,27 +61,12 @@ class Printer extends NetworkEquipment
             }
             unset($val->port);
 
-           //inventoried printers certainly have ethernet
-            if (property_exists($this->raw_data->content ?? new \stdClass(), 'network_device')) {
-                $val->have_ethernet = 1;
-            }
-
-           // Hack for USB Printer serial
+            // Hack for USB Printer serial
             if (
                 property_exists($val, 'serial')
                 && preg_match('/\/$/', $val->serial)
             ) {
                 $val->serial = preg_replace('/\/$/', '', $val->serial);
-            }
-
-            if (property_exists($val, 'ram')) {
-                $val->memory_size = $val->ram;
-                unset($val->ram);
-            }
-
-            if (property_exists($val, 'credentials')) {
-                $val->snmpcredentials_id = $val->credentials;
-                unset($val->credentials);
             }
 
             $res_rule = $rulecollection->processAllRules(['name' => $val->name]);
@@ -148,78 +82,18 @@ class Printer extends NetworkEquipment
                     $known_key = md5('manufacturers_id' . $res_rule['manufacturer']);
                     $this->known_links[$known_key] = $res_rule['manufacturer'];
                 }
-
-                if (isset($this->extra_data['pagecounters'])) {
-                    $pcounter = (object)$this->extra_data['pagecounters'];
-                    foreach ($mapping_pcounter as $origin => $dest) {
-                        if (property_exists($pcounter, $origin)) {
-                             $pcounter->$dest = $pcounter->$origin;
-                        }
-
-                        if (property_exists($pcounter, 'total_pages')) {
-                            $val->last_pages_counter = $pcounter->total_pages;
-                        }
-                        $this->counters = $pcounter;
-                    }
-                }
             } else {
                 unset($this->data[$k]);
             }
         }
-
-        //try to know if management port IP is already known as IP port
-        //if yes remove it from management port
-        $known_ports = $port_managment = $this->getManagementPorts();
-        if (isset($known_ports['management']) && property_exists($known_ports['management'], 'ipaddress')) {
-            foreach ($known_ports['management']->ipaddress as $pa_ip_key => $pa_ip_val) {
-                if (property_exists($this->raw_data->content, 'network_ports')) {
-                    foreach ($this->raw_data->content->network_ports as $port_obj) {
-                        if (property_exists($port_obj, 'ips')) {
-                            foreach ($port_obj->ips as $port_ip) {
-                                if ($pa_ip_val == $port_ip) {
-                                    unset($port_managment['management']->ipaddress[$pa_ip_key]);
-                                    if (empty($port_managment['management']->ipaddress)) {
-                                        unset($port_managment['management']);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        $this->setManagementPorts($port_managment);
 
         return $this->data;
     }
 
     public function handle()
     {
-        if ($this->item->getType() != GPrinter::getType()) {
-            if ($this->conf->import_printer == 1) {
-                $this->handleConnectedPrinter();
-            }
-            return;
-        }
-
-        parent::handle();
-        $this->handleMetrics();
-    }
-
-    /**
-     * Handle a printer connecter to a computer
-     *
-     * @return void
-     */
-    protected function handleConnectedPrinter()
-    {
         /** @var \DBmysql $DB */
         global $DB;
-
-        if (!$this->checkPrinterConf($this->conf)) {
-            return;
-        }
 
         $rule = new RuleImportAssetCollection();
         $printer = new GPrinter();
@@ -249,7 +123,7 @@ class Printer extends NetworkEquipment
                 $items_id = null;
                 $itemtype = \Printer::class;
                 if ($data['found_inventories'][0] == 0) {
-                   // add printer
+                    // add printer
                     $val->entities_id = $entities_id;
                     $val->is_recursive = $this->is_recursive;
                     $val->is_dynamic = 1;
@@ -308,7 +182,7 @@ class Printer extends NetworkEquipment
             $db_printers[$idtmp] = $data['id'];
         }
         if (count($db_printers)) {
-           // Check all fields from source:
+            // Check all fields from source:
             foreach ($printers as $key => $printers_id) {
                 foreach ($db_printers as $keydb => $prints_id) {
                     if ($printers_id == $prints_id) {
@@ -319,7 +193,7 @@ class Printer extends NetworkEquipment
                 }
             }
 
-           // Delete printers links in DB
+            // Delete printers links in DB
             foreach ($db_printers as $idtmp => $data) {
                 (new $lclass())->delete(['id' => $idtmp], true);
             }
@@ -338,81 +212,15 @@ class Printer extends NetworkEquipment
         }
     }
 
-    /**
-     * Get printer counters
-     *
-     * @return \stdClass
-     */
-    public function getCounters(): \stdClass
+    public function checkConf(Conf $conf): bool
     {
-        return $this->counters;
-    }
-
-    /**
-     * Handle printer metrics
-     *
-     * @return void
-     */
-    public function handleMetrics()
-    {
-        if ($this->counters === null) {
-            return;
-        }
-
-        $unicity_input = [
-            'printers_id' => $this->item->fields['id'],
-            'date'        => date('Y-m-d', strtotime($_SESSION['glpi_currenttime'])),
-        ];
-        $input = array_merge((array)$this->counters, $unicity_input);
-
-        $metrics = new PrinterLog();
-        if ($metrics->getFromDBByCrit($unicity_input)) {
-            $input['id'] = $metrics->fields['id'];
-            $metrics->update($input, false);
-        } else {
-            $metrics->add($input, [], false);
-        }
-    }
-
-    /**
-     * Try to know if printer need to be updated from discovery
-     * Only if IP has changed
-     * @return boolean
-     */
-    public static function needToBeUpdatedFromDiscovery(CommonDBTM $item, $val)
-    {
-        if (property_exists($val, 'ips')) {
-            foreach ($val->ips as $ip) {
-                $blacklist = new Blacklist();
-                //exclude IP if needed
-                if ('' != $blacklist->process(Blacklist::IP, $ip)) {
-                    //try to find IP (get from discovery) from known IP of Printer
-                    //if found refuse update
-                    //if no, printer IP have changed so  we allow the update from discovery
-                    $ipaddress = new IPAddress($ip);
-                    $tmp['mainitems_id'] = $item->fields['id'];
-                    $tmp['mainitemtype'] = $item::getType();
-                    $tmp['is_dynamic']   = 1;
-                    $tmp['name']         = $ipaddress->getTextual();
-                    if ($ipaddress->getFromDBByCrit($tmp)) {
-                        return false;
-                    }
-                    return true;
-                }
-            }
-        }
-        return false;
+        /** @var array $CFG_GLPI */
+        global $CFG_GLPI;
+        return $conf->import_printer == 1 && in_array($this->item::class, $CFG_GLPI['peripheralhost_types']);
     }
 
     public function getItemtype(): string
     {
         return \Printer::class;
-    }
-
-    public function checkPrinterConf(Conf $conf): bool
-    {
-        /** @var array $CFG_GLPI */
-        global $CFG_GLPI;
-        return $conf->import_printer == 1 && in_array($this->item::class, $CFG_GLPI['peripheralhost_types']);
     }
 }
